+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## 5.7.0
 
+### Features
+
+- [#983](https://github.com/okta/okta-auth-js/pull/983) Adds new method `setHeaders`
+
 ### Other
+
 - [#981](https://github.com/okta/okta-auth-js/pull/981) TypeScript: Allows optional paramters for IDX methods
 
 ## 5.6.0